Update per entity
