+++ src/gobupload/update/__init__.py
@@ -1,0 +1,1 @@
+from gobupload.update.main import full_update  # noqa: F401
